requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioplayer==0.6
blinker==1.9.0
Bootstrap-Flask==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dominate==2.9.1
elevenlabs==2.15.0
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
MarkupSafe==3.0.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
visitor==0.1.3
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
